Update SaeedMuhammad_assign3_problem2.py
Guess the Number Game
A program that lets the user guess an integer between 1 and 10. The integer is randomly selected by program when it starts up. Player is allowed to guess three times - if they guess correctly, they win and the program ends. If they guess incorrectly they are told if the number they guessed was too high or too low and then be asked to guess another time. At the end of the program it displays the secret number as well as how many attempts it took to guess that number.