Update SaeedMuhammad_assign4_part1.py
Roll the Dice until snake eyes 
A program which first prompts the user for the number of sides to the dice they will be rolling.  The program keeps "rolling the dice until it gets snake eyes. The program will "announce" every pair rolled and then tell the user how many "rolls" were required to get a pair of ones. In addition, it will calculate the average roll for each die and present how many times your roll came up with "doubles" along the way  to the user.